requirements.txt
openai==1.75.0
google-genai==1.24.0
langchain==0.3.24
langgraph==0.2.52
faiss-cpu==1.13.2
sentence-transformers==3.3.1
rank-bm25==0.2.2
tiktoken==0.8.0
yt-dlp==2025.3.31
PyPDF2==3.0.1
gitpython==3.1.43
requests==2.32.3
beautifulsoup4==4.12.3
redis==5.2.1
rich==13.9.4
python-dotenv==1.0.1
black==24.10.0
pyflakes==3.2.0
numpy==2.1.3
# asyncio is part of Python standard library and is not installed via pip.
